refactor: extract LegacyTaskBody from TasksListItem
Pure move ahead of adding a second task body. TasksListItem keeps the
accordion shell and expand/collapse; the free-response and upload UI
moves out unchanged.
